Add files via upload
I reached a major problem. I just realized/found out that you cannot invoke two paint() methods at the same time. I need to perform a major overhaul of my current design and change it dramatically. 

This version has clear problems in it, but the way that Thread is used is correct, hopefully.